Create separate classes and files. Change directory
